requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

sqlmodel==0.0.16
psycopg2-binary==2.9.9

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
bcrypt==4.0.1

requests==2.32.3
redis==5.0.4

pandas==2.2.2      
openpyxl==3.1.2       
fpdf2==2.7.8         

python-dotenv==1.0.1
pydantic[email]==2.7.1

loguru==0.7.2
